+++ src/bin/sks.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const FAST_PACKAGE_VERSION = '1.18.2';
+const FAST_PACKAGE_VERSION = '1.18.3';
 const args = process.argv.slice(2);
 
 try {